Reenable tent annexation
On Angel Island each player is given a 'rolled-up tent'. This tent can be placed anywhere a house can be placed.
In order to prevent players from griefing others by placing tents in all the premium housing locations, a player with a house can 'annex' that tent. Basically, when a player places a house over a tent, the tent is packed up and placed in the owners bankbox.
On the flip side, if you have to annex a player's tent to place your own house, you won't be able to drop or transfer your house for 7 days. This prevents players from griefing tent owners.